More Added Features
-aquired character builder which will allow us to let the player fully customize their character
-reorganized player prefabs
-reorganized folders within project

PROJECTS TO FINISH
-linking menu with character chooser and battlefield selector
-health bar reducing health on wrong player
-adding sword collision and damage amount
-creating default character models to work with during development
-fixing bug with background covering player in PirateShip scene
-eventually getting rid of the "The variable spawnPoint has not been assigned. Does not cause any issue but will eventually want to get rid of it.
-assigning spawnpoints and swapping them based on how many players are fighting.

I have added a lot and made a ton of changes but like usual if there is a complaint one of you want to change something, go right ahead. Especially when referring to the battlefields and menus. I found the pirate ship and the castle which I thought were cool so I added them for the time being. Also, this game folder is getting to be around 1 gb and will likely be larger as we advanced and add more and more so just be wary when pulling commit.