Big changes. this may be a long one.
So, had a major bug with the player movement so had to redo all of the characters and maps. No character has respawn or player health. Will have those working by tuesday. Made 4 players so if scripts need to be added to player, add to prefab from the folder view instead of clicking on the prefab in the scene.