Expanded level, attached respawn components to viking and wizard, changed friction on walls to be slippery.
Made a lot of changes in this one. I had to commit before proceeding, in case I changed too much and forgot. I played around with character values for mass, interpolation, jump forces, and the like. I moved the spawn bar much further to the top to create more jumping room. This was because I added small platforms to the level to expand level playability (the spawn bar was blocking the jumps). I enlarged the spaces between the main platforms so the viking could actually fall through. The viking and wizard have respawn components attached so now they die and respawn accordingly.