Even more gameplay tweaks
Finally got mass, force, speed, jump values to where it feels good (imo). Had problems with characters getting caught on the sides of the platforms. This time because of the platform caps, so I applied an ugly solution of squeezing the platform slightly and lowering it to basically meet the level of the pillars.

Note: Viking kind of a pain in the ass to design the level around because of his way larger size. Might want to replace with high-speed thief that can shank archer and wizard.