Animations...animations..animations
Finished the running, attacking, and idle animations for wizard, archer, and knight.
Changed the pirate level camera to fit screen.
Added in orginal sprites for use with animations and prefabs into characterPrefabs.
Organized folders.

TO DO:
Fix player jump animations for the animated sprite raises y axis of image of character but not character. \
Add in damage that will come from sword and other melee attacks.
Add in theif and viking animations.