In the Description
Fully functioning Spawning systemwhere multiple players can be the same character.

All data for project that will be needed globally is stored in playerValues.

Character selection is unfavorable but it was the only way I could figure it out. It does work with 4 controllers but I was having issues with button input on both controllers but not horiztontal.

The Viking is acting up overall so we will need to check him in the morning.

BIG BUGS TO LOOK AT
-------------------------
Make sure player colliders are not getting caught on walls.
Fixing menu selection. Honestly think we should ditch the other two levels due to time and focus our time on making two really good maps.

Player input needs to work with shooting and attacking as well as jumping.

First thing to hit though is the scene navigation. From the last character selection scene I can't call the scene the user originally chose. Let me know if you have any questions and I will see you tomorrow.